Pls i want ya

+++ Client/SocketInOut.cpp
@@ -71,7 +71,7 @@ void SocketInOut::InitSockIn() {
 
 void SocketInOut::Send(const string& msg){
 	string str = msg + "\r\n";
-	if (send(sock, str.c_str(), str.size(), 0) == SOCKET_ERROR)
+	if (send(sock, str.c_str(), (int)str.size(), 0) == SOCKET_ERROR)
 		throw Erreur(WSAGetLastError(), "Error can not send message");
 }
